+++ modules/meteroid/crates/diesel-models/src/query/coupons.rs
@@ -1,7 +1,7 @@
 use crate::coupons::{CouponRow, CouponRowNew, CouponRowPatch};
 use crate::errors::IntoDbResult;
 use crate::{DbResult, PgConn};
-use diesel::{debug_query, ExpressionMethods, QueryDsl};
+use diesel::{debug_query, ExpressionMethods, QueryDsl, SelectableHelper};
 use diesel_async::RunQueryDsl;
 use error_stack::ResultExt;
 use tap::TapFallible;
@@ -98,6 +98,31 @@ impl CouponRow {
             .await
             .tap_err(|e| log::error!("Error while fetching coupons: {:?}", e))
             .attach_printable("Error while fetching coupons")
+            .into_db_result()
+    }
+
+    pub async fn list_by_subscription_id(
+        conn: &mut PgConn,
+        tenant_id: &uuid::Uuid,
+        subscription_id: &uuid::Uuid,
+    ) -> DbResult<Vec<CouponRow>> {
+        use crate::schema::coupon::dsl as c_dsl;
+        use crate::schema::subscription::dsl as s_dsl;
+        use crate::schema::subscription_coupon::dsl as sc_dsl;
+
+        let query = sc_dsl::subscription_coupon
+            .inner_join(s_dsl::subscription)
+            .inner_join(c_dsl::coupon)
+            .filter(sc_dsl::subscription_id.eq(subscription_id))
+            .filter(s_dsl::tenant_id.eq(tenant_id))
+            .select(CouponRow::as_select());
+
+        log::debug!("{}", debug_query::<diesel::pg::Pg, _>(&query).to_string());
+
+        query
+            .get_results(conn)
+            .await
+            .attach_printable("Error while listing Coupon by subscription_id")
             .into_db_result()
     }
 }
