+++ modules/meteroid/crates/diesel-models/src/subscription_coupons.rs
@@ -9,13 +9,6 @@ pub struct SubscriptionCouponRow {
     pub id: Uuid,
     pub subscription_id: Uuid,
     pub coupon_id: Uuid,
-    pub coupon_code: String,
-    pub coupon_description: String,
-    pub coupon_discount: serde_json::Value,
-    pub coupon_expires_at: Option<NaiveDateTime>,
-    pub coupon_redemption_limit: Option<i32>,
-    pub coupon_recurring_value: i32,
-    pub coupon_reusable: bool,
     pub created_at: NaiveDateTime,
 }
 
@@ -26,11 +19,4 @@ pub struct SubscriptionCouponRowNew {
     pub id: Uuid,
     pub subscription_id: Uuid,
     pub coupon_id: Uuid,
-    pub coupon_code: String,
-    pub coupon_description: String,
-    pub coupon_discount: serde_json::Value,
-    pub coupon_expires_at: Option<NaiveDateTime>,
-    pub coupon_redemption_limit: Option<i32>,
-    pub coupon_recurring_value: i32,
-    pub coupon_reusable: bool,
 }
