+++ modules/meteroid/crates/meteroid-store/src/domain/subscriptions.rs
@@ -3,11 +3,12 @@ use chrono::{NaiveDate, NaiveDateTime};
 use o2o::o2o;
 use uuid::Uuid;
 
+use crate::domain::coupons::Coupon;
 use crate::domain::enums::BillingPeriodEnum;
 use crate::domain::subscription_add_ons::{CreateSubscriptionAddOns, SubscriptionAddOn};
 use crate::domain::{
     BillableMetric, CreateSubscriptionComponents, CreateSubscriptionCoupons, Schedule,
-    SubscriptionComponent, SubscriptionCoupon,
+    SubscriptionComponent,
 };
 use diesel_models::subscriptions::SubscriptionRowNew;
 use diesel_models::subscriptions::{
@@ -171,7 +172,7 @@ pub struct SubscriptionDetails {
     pub schedules: Vec<Schedule>,
     pub price_components: Vec<SubscriptionComponent>,
     pub add_ons: Vec<SubscriptionAddOn>,
-    pub coupons: Vec<SubscriptionCoupon>,
+    pub coupons: Vec<Coupon>,
     pub metrics: Vec<BillableMetric>,
     pub mrr_cents: u64,
